+++ include/tempo.h
@@ -1,0 +1,7 @@
+#ifndef TEMPO_H
+#define TEMPO_H
+
+// Protótipos das função 
+void tempo();
+
+#endif // AREA_H
